feat(#96): Add SummaryStats component for dashboard
Made-with: Cursor
